Set vm name on iscsi and adm
Recently I tried halting my vms and bringing them back up again and
noticed that I wasn't able to do so because it could not locate the
iscsi vm based on the hostname. This meant that it could no longer check
to see if the SATA controller exists on the iscsi vm and causes the
"vagrant up iscsi" command to fail. Setting the vbox name to "iscsi" and
then searching for the sata controller using the vbox name instead of
the vm's hostname fixes the problem. I updated the vbox name for the
admin node as well. The mds's and oss's already have their vbox names
set.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>