When developing with docker locally there are some network (#84)
configurations that need to be made to each of the storage server nodes
before they can be added to IML in the docker container. The id_rsa file
is already copied from local ~/.ssh to the job scheduler container.
However, the local public key needs to be copied into the authorized
keys file for each storage server node. This will allow the docker
container to add the storage servers using the existing key feature in
IML.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>